requirements.txt
streamlit==1.11.1
stqdm>=0.0.2
streamlit-pydantic==0.5.0
streamlit-authenticator==0.1.4
notebook>=6.4.1
PyYAML==5.4
pandas>=1.3.5
openpyxl==3.0.7

# mkt data packages
requests>=2.28.2
urllib3>=1.26.14
yfinance>=0.2.12
yahooquery>=2.3.0
requests-cache==0.6.3
businessdate==0.5

# plotting
plotly==4.14.3
